upd files

+++ insight_stream/client.py
@@ -2,24 +2,35 @@ from typing import Dict, List, Optional
 import json
 import os
 from langchain_openai import OpenAIEmbeddings
-from langchain_community.vectorstores.qdrant import Qdrant
+from langchain_qdrant import Qdrant
 from qdrant_client import QdrantClient, models
 from langchain_core.documents import Document
 from langchain_community.document_loaders import UnstructuredFileLoader
-from langchain.text_splitter import TokenTextSplitter
 import requests
 import urllib.parse
+from insight_stream.embed import CustomEmbeddings
+import socket
 
-TOKEN = os.getenv('TOKEN')
+OPENAI_BASE = os.getenv('OPENAI_BASE')
 OPENAI_API_KEY = os.getenv('OPENAI_API_KEY')
+EMBED_MODEL = os.getenv('EMBED_MODEL')
+
 QDRANT_URL = os.getenv('QDRANT_URL')
 QDRANT_KEY = os.getenv('QDRANT_KEY')
-OPENAI_API_BASE = os.getenv('OPENAI_API_BASE')
-SERVER_NAME = os.getenv('SERVER_NAME')
+QDRANT_VECTOR_SIZE = int(os.getenv('QDRANT_VECTOR_SIZE', 1536))
 
-embedding = OpenAIEmbeddings(model='text-embedding-ada-002', 
+SERVER_NAME = os.getenv('SERVER_NAME')
+TOKEN = os.getenv('TOKEN')
+
+embedding = CustomEmbeddings()
+embedding.base_url = f"{OPENAI_BASE}/embeddings"
+embedding.model = EMBED_MODEL
+
+"""embedding = OpenAIEmbeddings(model=EMBED_MODEL,
 							 api_key=OPENAI_API_KEY, 
-							 openai_api_base=OPENAI_API_BASE)
+							 openai_api_base=OPENAI_BASE),
+							 model_kwargs={"encoding_format": "float"}"""
+
 clientQdrant=QdrantClient(url=QDRANT_URL, api_key=QDRANT_KEY)
 
 def ask(
@@ -27,9 +38,8 @@ def ask(
 		question: str
 	) -> Optional[Dict[str, any]]:
 	""""Получение ответа бота на вопрос пользователя"""
-	
 	request = requests.post(
-		f'https://{SERVER_NAME}/v.1.0/{bot_id}', 
+		f'{SERVER_NAME}/v.1.0/{bot_id}',
 		data=json.dumps({ "question": question }),
 		headers={
 			'Content-Type': 'application/json',
@@ -41,12 +51,12 @@ def ask(
 	else:
 		return None
 	
-def _add_documents(index_id: str, documents: List[Document], file_path: str = 'file') -> List[Document]: #значения по умолчанию для folder_name и file_path
+def _add_documents(index_id: str, documents: List[Document], file_path: str) -> List[Document]:
 	"""Добавляем в квадрант документы, являющиеся чанками для одного файла из папки"""
 	try:
 		clientQdrant.create_collection(
 				collection_name=index_id,
-				vectors_config=models.VectorParams(size=1536, distance=models.Distance.COSINE)
+				vectors_config=models.VectorParams(size=QDRANT_VECTOR_SIZE, distance=models.Distance.COSINE)
 			)
 	except:
 		pass
@@ -56,10 +66,13 @@ def _add_documents(index_id: str, documents: List[Document], file_path: str = 'f
 	for doc in documents:
 		new_doc = Document(page_content=doc.page_content, metadata = {
 			"topic": os.path.splitext(os.path.basename(file_path))[0], 
-			"url": urllib.parse.quote(f'https://{SERVER_NAME}/documents/{os.path.basename(file_path)}', safe=":/")})
+			"url": urllib.parse.quote(f'{SERVER_NAME}/documents/{os.path.basename(file_path)}', safe=":/")})
 		docs_for_qdrant.append(new_doc)
-
-	qdrant.add_documents(docs_for_qdrant)
+		
+	try:
+		qdrant.add_documents(docs_for_qdrant)
+	except Exception as e:
+		print(e)
 	return docs_for_qdrant
 
 
@@ -80,44 +93,54 @@ def delete_documents(index_id: str, documents: List[Document]):
 		_del_file_from_server(url)
 
 
-def upload_doc(index_id: str, path: str):
+def upload_doc(index_id: str, path: str) -> List[Document]:
 	"""Загрузка документов в индекс квадранта и на сервер"""
 
 	loader = UnstructuredFileLoader(path)
-
-	token_text_splitter = TokenTextSplitter(         
-		model_name='gpt-3.5-turbo',
-		chunk_size = 512, 
-		chunk_overlap = 50
-	)
-
-	chunks = loader.load_and_split(text_splitter=token_text_splitter)
+	chunks = loader.load_and_split()
 
 	#загрузка чанков документа в квадрант
 	docs = _add_documents(index_id, chunks, path)
-	print(f"Документы загружены в квадрант, index_id {index_id}")
+	print(f"Файл {os.path.basename(path)} загружен в квадрант, index_id {index_id}")
 
 	#загрузка файла на сервер
 	url = _load_file_to_server(path)
 	
 	return docs
 
-def upload_dir(index_id: str, path: str):
-	"""Множественная загрузка документов в индекс квадранта и на  сервер"""
+def upload_dir(index_id: str, path: str) -> List[Document]:
+	"""Множественная загрузка документов в индекс квадранта и на сервер"""
 
-	#загрузка файлов на сервер
-	urls = _load_dir_to_server(path)
+	docs = []
 
 	for file in os.listdir(path):
 		file_path = os.path.join(path, file)
 		if os.path.isfile(file_path):
 			chunks = upload_doc(index_id, file_path)
-			yield from chunks
+			docs.extend(chunks)
+
+	return docs
+
+def _check_server_availability() -> bool:
+	"Проверка доступности сервера"
+
+	server_url = urllib.parse.urlparse(SERVER_NAME)
+	host = server_url.hostname
+	port = server_url.port
+	try:
+		socket.create_connection((host, port), timeout=5)
+		return True
+	except socket.error:
+		return False
 
 def _load_file_to_server(file_path: str) -> str:
 	"""Добавление файла на сервера"""
 
-	url = f"https://{SERVER_NAME}/documents/{os.path.basename(file_path)}"
+	if not _check_server_availability():
+		print("Сервер для загрузки документов недоступен")
+		return ''
+
+	url = f"{SERVER_NAME}/documents/{os.path.basename(file_path)}"
 	file_url = urllib.parse.quote(url, safe=":/")
 
 	headers = {'Authorization': f'Bearer {TOKEN}'}
@@ -126,14 +149,18 @@ def _load_file_to_server(file_path: str) -> str:
 	response = requests.put(url, headers=headers, files=files)
     
 	if response.status_code == 201:
-		print(f"Успешно загружен файл {os.path.basename(file_path)}!")
+		print(f"Успешно загружен файл {os.path.basename(file_path)} на сервер!")
 		return file_url
 	else:
-		print("Загрузка не удалась")
+		print(f"Загрузка файла {os.path.basename(file_path)} на сервер не удалась")
 		return ''
 	
 def _del_file_from_server(url: str):
 	"""Удаление файла с сервера"""
+
+	if not _check_server_availability():
+		print("Сервер недоступен")
+		return ''
 
 	headers = {'Authorization': f'Bearer {TOKEN}'}
 	response = requests.delete(url, headers=headers)
@@ -142,19 +169,3 @@ def _del_file_from_server(url: str):
 		print(f"Успешно удален файл {url}!")
 	else:
 		print("Ошибка при удалении")
-	
-
-def _load_dir_to_server(path: str) -> List[str]:
-	"""Множественная загрузка файлов на сервер"""
-
-	urls = []
-    
-	if os.path.isfile(path):
-		urls.append(_load_file_to_server(path))
-	elif os.path.isdir(path):
-		for file in os.listdir(path):
-			file_path = os.path.join(path, file)
-			if os.path.isfile(file_path):
-				urls.append(_load_file_to_server(file_path))
-    
-	return urls
